pyproject.toml
[project]
name = "ai-health-agent"
version = "0.1.0"
description = "Health management AI agent scaffold using OpenAI LLM"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai>=1.40.0",
  "python-dotenv>=1.0.1",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"